Add this month's game - better later than nothing
This is the remake of the very first game I ever wrote, in QBasic, when I was 11. It was called "Jogando dados com Duke Nukem" ("playing dice with Duke Nukem", in a very rough translation). I decided to avoid copyright infringement this time.

I had a very chaotic month, with an injury that got me bed ridden and some work travelling, but I had a spare hour in my hands since I had to wait for an appointment. I came precisely one hour earlier than I was supposed to - it felt like the perfect opportunity.